src/warnerco/backend/pyproject.toml
[project]
name = "warnerco-robotics-schematica"
version = "1.0.0"
description = "WARNERCO Robotics Schematica - Agentic robot schematics system with semantic memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "fastmcp>=2.3.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "chromadb>=0.5.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.17",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
azure = [
    "azure-search-documents>=11.5.0",
    "azure-identity>=1.19.0",
    "azure-core>=1.32.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
# MCP stdio server (for Claude Desktop, VS Code, etc.)
warnerco-mcp = "app.mcp_stdio:main"
# HTTP server
warnerco-serve = "app.main:run_server"

# Poetry compatibility (reads from [project] above)
[tool.poetry]
name = "warnerco-robotics-schematica"
version = "1.0.0"
description = "WARNERCO Robotics Schematica - Agentic robot schematics system with semantic memory"
authors = ["WARNERCO <engineering@warnerco.io>"]
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
fastmcp = "^2.3.0"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.3.0"
chromadb = "^0.5.0"
httpx = "^0.28.0"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.0"

[tool.poetry.group.azure.dependencies]
azure-search-documents = "^11.5.0"
azure-identity = "^1.19.0"
azure-core = "^1.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"

[tool.poetry.scripts]
warnerco-mcp = "app.mcp_stdio:main"
warnerco-serve = "app.main:run_server"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# uv compatibility
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
